Use Wayback Machine for the the CH340 driver link
I have observed over a long period of time that the wch.cn website is no longer online. I also have seen wch-ic.com, which SparkFun says is the English version of the site:
https://learn.sparkfun.com/tutorials/how-to-install-ch340-drivers/all#drivers-if-you-need-them
but that site is down too.

SparkFun does provide driver downloads on that page, but unfortunately they are the outdated 3.4 instead of the most recent 3.5 version (I only checked for Windows).

So I think Wayback Machine is the best option. Unfortunately, there is something about the WCH website that makes the download links from the driver download page not work:
https://web.archive.org/web/20210127011437/http://www.wch.cn/download/CH341SER_EXE.html
so the only option is to provide a link directly to the archived driver installer.